refactor(cli): consolidate the gateway into `micro gateway`
Rename `micro server` to `micro gateway` and fold the standalone
micro-mcp-gateway binary into it, so there is one gateway with one name.

- Rename command server -> gateway (package cmd/micro/server ->
  cmd/micro/gateway; update the run and main importers).
- `micro gateway` gains the MCP production controls the standalone binary
  had: --rate-limit/--rate-burst, --auth, --audit, --scope,
  --circuit-breaker[-timeout], and --x402-*. When --mcp-address is set it
  runs the full gateway/mcp with those options; registry selection uses
  the CLI global --registry/--registry_address flags.
- `micro server` stays as a hidden, deprecated alias that warns and
  delegates, so existing invocations keep working.
- Delete cmd/micro-mcp-gateway (binary + Dockerfile). goreleaser only
  builds ./cmd/micro, so releases are unaffected.
- Repoint deploy/refs at the micro CLI image: Dockerfile CMD -> gateway;
  Helm chart runs `micro gateway --mcp-address` with registry via
  MICRO_REGISTRY env, image ghcr.io/micro/go-micro; drop the v6
  micro-mcp-gateway vanity entry and rename server -> gateway; update the
  docs guides and CLAUDE.md.

Build, vet, gofmt, and the run/gateway/cmd tests pass; `micro gateway`
and the hidden `micro server` alias both resolve.
