+++ cmd/micro/gateway/gateway.go
@@ -1,4 +1,4 @@
-package server
+package gateway
 
 import (
 	"fmt"
