+++ cmd/micro/gateway/util_jwt.go
@@ -1,4 +1,4 @@
-package server
+package gateway
 
 import (
 	"crypto/rand"
